+++ src/sat/utils/geo/Coordinates.java
@@ -1,31 +1,32 @@
 package sat.utils.geo;
 
 public class Coordinates {
-	private int x;
-	private int y;
-	private int z;
+	private float x;
+	private float y;
+	private float z;
 
-	public Coordinates(int x, int y, int z) {
+	public Coordinates(float x, float y, float z) {
 		this.x = x;
 		this.y = y;
 		this.z = z;
 	}
-	
-	public int getX() {
+
+	public float getX() {
 		return x;
 	}
-	
-	public int getY() {
+
+	public float getY() {
 		return y;
 	}
-	
-	public int getZ() {
+
+	public float getZ() {
 		return z;
 	}
-	
+
 	public boolean equals(Coordinates obj) {
 		if(obj.getX() == this.x && obj.getY() == this.y && obj.getZ() == this.z)
 			return true;
+
 		return false;
 	}
 }
